refactor: SideBar 컴포넌트 - CalendarList 컴포넌트 재사용 및 폴더 구조 변경 (#28)
